refactor(#890): extract AmberCockpit component (behavior-preserving)
B-PR1 of the LCD twin-skin epic (#887). Pure move — the entire
current AmberLcdDisplay body (script, template, styles) becomes
AmberCockpit.svelte. AmberLcdDisplay stays as a thin delegate
for backwards compatibility (MobileRadioLayout, LcdSkin, tests).

LcdLayout now accepts a `variant: 'cockpit' | 'scope'` prop; the
default preserves today's behavior. The `scope` variant falls
through to AmberCockpit until C-PR1 (#895) delivers a dedicated
AmberScope component.

Closes #890
Part of #887
